// src/BindingChaos.Web/src/api/models/TrendPointResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BindingChaos.Web.Gateway
 * BindingChaos Web Gateway API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Generic response model for a single data point in any trend.
 * @export
 * @interface TrendPointResponse
 */
export interface TrendPointResponse {
    /**
     * Gets the date and time of this trend point in ISO 8601 format.
     * @type {string}
     * @memberof TrendPointResponse
     */
    date?: string;
    /**
     * Gets the type of event that occurred at this point in time.
     * The specific values depend on the context (e.g., "support", "oppose", "amplify", "attenuate").
     * @type {string}
     * @memberof TrendPointResponse
     */
    eventType?: string;
}

/**
 * Check if a given object implements the TrendPointResponse interface.
 */
export function instanceOfTrendPointResponse(value: object): value is TrendPointResponse {
    return true;
}

export function TrendPointResponseFromJSON(json: any): TrendPointResponse {
    return TrendPointResponseFromJSONTyped(json, false);
}

export function TrendPointResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): TrendPointResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'date': json['date'] == null ? undefined : json['date'],
        'eventType': json['eventType'] == null ? undefined : json['eventType'],
    };
}

export function TrendPointResponseToJSON(json: any): TrendPointResponse {
    return TrendPointResponseToJSONTyped(json, false);
}

export function TrendPointResponseToJSONTyped(value?: TrendPointResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'date': value['date'],
        'eventType': value['eventType'],
    };
}